requirements.txt
anyio==4.3.0
db_query==1.3.0
DBUtils==3.1.0
pika==1.3.2
PyMySQL==1.1.1
python-dotenv==1.0.1
Requests==2.32.3
starlette==0.38.5
websockets==13.0.1
